Initial commit, covers through Patch 0727.
Updates from here on out will require an updated DynamicTranslationLoader that should be coming out in the next few days, presumably to be released in BepisPlugins R5.

There will also be an update due to the heavy graphical elements contained in the Studio, with edited image/TextureAtlas support. This will also be forthcoming in the updated BepisPlugins R5 release.