Added Afterschool Preorder DLCs.
Thanks to the Anon who dug those up and got them to us. Note as always that this will not add the content to the game if you lack it - merely translate it. You'll still need to hunt the actual DLC content down if you want it in-game.

Also took the opportunity to fix a few small mistakes/leftovers from the massive update.